@@ -1,23 +1,20 @@
 module clog
 
-go 1.23.0
-
-toolchain go1.24.2
+go 1.25.1
 
 require (
 	github.com/ClickHouse/clickhouse-go/v2 v2.34.0
-	github.com/andypangaribuan/gmod v1.0.48
-	github.com/questdb/go-questdb-client/v3 v3.2.0
-	github.com/stretchr/testify v1.10.0
-	google.golang.org/grpc v1.65.0
-	google.golang.org/protobuf v1.34.2
+	github.com/andypangaribuan/gmod v1.0.74
+	github.com/stretchr/testify v1.11.1
+	google.golang.org/grpc v1.77.0
+	google.golang.org/protobuf v1.36.10
 )
 
 require (
 	cloud.google.com/go v0.115.1 // indirect
 	cloud.google.com/go/auth v0.9.0 // indirect
 	cloud.google.com/go/auth/oauth2adapt v0.2.4 // indirect
-	cloud.google.com/go/compute/metadata v0.5.0 // indirect
+	cloud.google.com/go/compute/metadata v0.9.0 // indirect
 	cloud.google.com/go/iam v1.1.13 // indirect
 	cloud.google.com/go/storage v1.43.0 // indirect
 	github.com/ClickHouse/ch-go v0.66.0 // indirect
