@@ -8,8 +8,8 @@
 
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.30.0
-// 	protoc        v3.21.12
+// 	protoc-gen-go v1.36.10
+// 	protoc        v5.29.3
 // source: res/proto/clog.proto
 
 package sclog
@@ -20,6 +20,7 @@ import (
 	wrapperspb "google.golang.org/protobuf/types/known/wrapperspb"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -31,21 +32,18 @@ const (
 
 // status: enum [ok, !ok]
 type Response struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	Status        string                 `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
+	Message       string                 `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Status  string `protobuf:"bytes,1,opt,name=status,proto3" json:"status,omitempty"`
-	Message string `protobuf:"bytes,2,opt,name=message,proto3" json:"message,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *Response) Reset() {
 	*x = Response{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *Response) String() string {
@@ -56,7 +54,7 @@ func (*Response) ProtoMessage() {}
 
 func (x *Response) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -85,30 +83,29 @@ func (x *Response) GetMessage() string {
 	return ""
 }
 
+// occurredAt : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 type RequestNoteV1 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState  `protogen:"open.v1"`
+	Uid           string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
+	UserId        *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
+	PartnerId     *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
+	SvcName       string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
+	SvcVersion    string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
+	ExecPath      string                  `protobuf:"bytes,6,opt,name=execPath,proto3" json:"execPath,omitempty"`
+	ExecFunction  string                  `protobuf:"bytes,7,opt,name=execFunction,proto3" json:"execFunction,omitempty"`
+	Key           *wrapperspb.StringValue `protobuf:"bytes,8,opt,name=key,proto3" json:"key,omitempty"`
+	SubKey        *wrapperspb.StringValue `protobuf:"bytes,9,opt,name=subKey,proto3" json:"subKey,omitempty"`
+	Data          string                  `protobuf:"bytes,10,opt,name=data,proto3" json:"data,omitempty"`
+	OccurredAt    string                  `protobuf:"bytes,11,opt,name=occurredAt,proto3" json:"occurredAt,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Uid          string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
-	UserId       *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
-	PartnerId    *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
-	SvcName      string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
-	SvcVersion   string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
-	ExecPath     string                  `protobuf:"bytes,6,opt,name=execPath,proto3" json:"execPath,omitempty"`
-	ExecFunction string                  `protobuf:"bytes,7,opt,name=execFunction,proto3" json:"execFunction,omitempty"`
-	Key          *wrapperspb.StringValue `protobuf:"bytes,8,opt,name=key,proto3" json:"key,omitempty"`
-	SubKey       *wrapperspb.StringValue `protobuf:"bytes,9,opt,name=subKey,proto3" json:"subKey,omitempty"`
-	Data         string                  `protobuf:"bytes,10,opt,name=data,proto3" json:"data,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RequestNoteV1) Reset() {
 	*x = RequestNoteV1{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestNoteV1) String() string {
@@ -119,7 +116,7 @@ func (*RequestNoteV1) ProtoMessage() {}
 
 func (x *RequestNoteV1) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -204,41 +201,47 @@ func (x *RequestNoteV1) GetData() string {
 	return ""
 }
 
+func (x *RequestNoteV1) GetOccurredAt() string {
+	if x != nil {
+		return x.OccurredAt
+	}
+	return ""
+}
+
 // severity   : enum [success, warning, error]
 // startedAt  : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 // finishedAt : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 type RequestDbqV1 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState  `protogen:"open.v1"`
+	Uid           string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
+	UserId        *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
+	PartnerId     *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
+	SvcName       string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
+	SvcVersion    string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
+	SqlQuery      string                  `protobuf:"bytes,6,opt,name=sqlQuery,proto3" json:"sqlQuery,omitempty"`
+	SqlArgs       *wrapperspb.StringValue `protobuf:"bytes,7,opt,name=sqlArgs,proto3" json:"sqlArgs,omitempty"`
+	Severity      string                  `protobuf:"bytes,8,opt,name=severity,proto3" json:"severity,omitempty"`
+	ExecPath      string                  `protobuf:"bytes,9,opt,name=execPath,proto3" json:"execPath,omitempty"`
+	ExecFunction  string                  `protobuf:"bytes,10,opt,name=execFunction,proto3" json:"execFunction,omitempty"`
+	ErrorMessage  *wrapperspb.StringValue `protobuf:"bytes,11,opt,name=errorMessage,proto3" json:"errorMessage,omitempty"`
+	StackTrace    *wrapperspb.StringValue `protobuf:"bytes,12,opt,name=stackTrace,proto3" json:"stackTrace,omitempty"`
+	DbName        *wrapperspb.StringValue `protobuf:"bytes,13,opt,name=dbName,proto3" json:"dbName,omitempty"`
+	SchemaName    *wrapperspb.StringValue `protobuf:"bytes,14,opt,name=schemaName,proto3" json:"schemaName,omitempty"`
+	Host1         string                  `protobuf:"bytes,15,opt,name=host1,proto3" json:"host1,omitempty"`
+	Host2         *wrapperspb.StringValue `protobuf:"bytes,16,opt,name=host2,proto3" json:"host2,omitempty"`
+	Duration1     int32                   `protobuf:"varint,17,opt,name=duration1,proto3" json:"duration1,omitempty"`
+	Duration2     *wrapperspb.Int32Value  `protobuf:"bytes,18,opt,name=duration2,proto3" json:"duration2,omitempty"`
+	StartedAt     string                  `protobuf:"bytes,19,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
+	FinishedAt    string                  `protobuf:"bytes,20,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Uid          string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
-	UserId       *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
-	PartnerId    *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
-	SvcName      string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
-	SvcVersion   string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
-	SqlQuery     string                  `protobuf:"bytes,6,opt,name=sqlQuery,proto3" json:"sqlQuery,omitempty"`
-	SqlArgs      *wrapperspb.StringValue `protobuf:"bytes,7,opt,name=sqlArgs,proto3" json:"sqlArgs,omitempty"`
-	Severity     string                  `protobuf:"bytes,8,opt,name=severity,proto3" json:"severity,omitempty"`
-	ExecPath     string                  `protobuf:"bytes,9,opt,name=execPath,proto3" json:"execPath,omitempty"`
-	ExecFunction string                  `protobuf:"bytes,10,opt,name=execFunction,proto3" json:"execFunction,omitempty"`
-	ErrorMessage *wrapperspb.StringValue `protobuf:"bytes,11,opt,name=errorMessage,proto3" json:"errorMessage,omitempty"`
-	StackTrace   *wrapperspb.StringValue `protobuf:"bytes,12,opt,name=stackTrace,proto3" json:"stackTrace,omitempty"`
-	Host1        string                  `protobuf:"bytes,13,opt,name=host1,proto3" json:"host1,omitempty"`
-	Host2        *wrapperspb.StringValue `protobuf:"bytes,14,opt,name=host2,proto3" json:"host2,omitempty"`
-	Duration1    int32                   `protobuf:"varint,15,opt,name=duration1,proto3" json:"duration1,omitempty"`
-	Duration2    *wrapperspb.Int32Value  `protobuf:"bytes,16,opt,name=duration2,proto3" json:"duration2,omitempty"`
-	StartedAt    string                  `protobuf:"bytes,17,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
-	FinishedAt   string                  `protobuf:"bytes,18,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RequestDbqV1) Reset() {
 	*x = RequestDbqV1{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[2]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[2]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestDbqV1) String() string {
@@ -249,7 +252,7 @@ func (*RequestDbqV1) ProtoMessage() {}
 
 func (x *RequestDbqV1) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[2]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -348,6 +351,20 @@ func (x *RequestDbqV1) GetStackTrace() *wrapperspb.StringValue {
 	return nil
 }
 
+func (x *RequestDbqV1) GetDbName() *wrapperspb.StringValue {
+	if x != nil {
+		return x.DbName
+	}
+	return nil
+}
+
+func (x *RequestDbqV1) GetSchemaName() *wrapperspb.StringValue {
+	if x != nil {
+		return x.SchemaName
+	}
+	return nil
+}
+
 func (x *RequestDbqV1) GetHost1() string {
 	if x != nil {
 		return x.Host1
@@ -393,38 +410,35 @@ func (x *RequestDbqV1) GetFinishedAt() string {
 // startedAt  : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 // finishedAt : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 type RequestHttpCallV1 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState  `protogen:"open.v1"`
+	Uid           string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
+	UserId        *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
+	PartnerId     *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
+	SvcName       string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
+	SvcVersion    string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
+	Url           string                  `protobuf:"bytes,6,opt,name=url,proto3" json:"url,omitempty"`
+	Severity      string                  `protobuf:"bytes,7,opt,name=severity,proto3" json:"severity,omitempty"`
+	ReqHeader     *wrapperspb.StringValue `protobuf:"bytes,8,opt,name=reqHeader,proto3" json:"reqHeader,omitempty"`
+	ReqParam      *wrapperspb.StringValue `protobuf:"bytes,9,opt,name=reqParam,proto3" json:"reqParam,omitempty"`
+	ReqQuery      *wrapperspb.StringValue `protobuf:"bytes,10,opt,name=reqQuery,proto3" json:"reqQuery,omitempty"`
+	ReqForm       *wrapperspb.StringValue `protobuf:"bytes,11,opt,name=reqForm,proto3" json:"reqForm,omitempty"`
+	ReqFiles      *wrapperspb.StringValue `protobuf:"bytes,12,opt,name=reqFiles,proto3" json:"reqFiles,omitempty"`
+	ReqBody       *wrapperspb.StringValue `protobuf:"bytes,13,opt,name=reqBody,proto3" json:"reqBody,omitempty"`
+	ResData       *wrapperspb.StringValue `protobuf:"bytes,14,opt,name=resData,proto3" json:"resData,omitempty"`
+	ResCode       int32                   `protobuf:"varint,15,opt,name=resCode,proto3" json:"resCode,omitempty"`
+	ErrMessage    *wrapperspb.StringValue `protobuf:"bytes,16,opt,name=errMessage,proto3" json:"errMessage,omitempty"`
+	StackTrace    *wrapperspb.StringValue `protobuf:"bytes,17,opt,name=stackTrace,proto3" json:"stackTrace,omitempty"`
+	StartedAt     string                  `protobuf:"bytes,18,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
+	FinishedAt    string                  `protobuf:"bytes,19,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	Uid        string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
-	UserId     *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
-	PartnerId  *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
-	SvcName    string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
-	SvcVersion string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
-	Url        string                  `protobuf:"bytes,6,opt,name=url,proto3" json:"url,omitempty"`
-	Severity   string                  `protobuf:"bytes,7,opt,name=severity,proto3" json:"severity,omitempty"`
-	ReqHeader  *wrapperspb.StringValue `protobuf:"bytes,8,opt,name=reqHeader,proto3" json:"reqHeader,omitempty"`
-	ReqParam   *wrapperspb.StringValue `protobuf:"bytes,9,opt,name=reqParam,proto3" json:"reqParam,omitempty"`
-	ReqQuery   *wrapperspb.StringValue `protobuf:"bytes,10,opt,name=reqQuery,proto3" json:"reqQuery,omitempty"`
-	ReqForm    *wrapperspb.StringValue `protobuf:"bytes,11,opt,name=reqForm,proto3" json:"reqForm,omitempty"`
-	ReqFiles   *wrapperspb.StringValue `protobuf:"bytes,12,opt,name=reqFiles,proto3" json:"reqFiles,omitempty"`
-	ReqBody    *wrapperspb.StringValue `protobuf:"bytes,13,opt,name=reqBody,proto3" json:"reqBody,omitempty"`
-	ResData    *wrapperspb.StringValue `protobuf:"bytes,14,opt,name=resData,proto3" json:"resData,omitempty"`
-	ResCode    int32                   `protobuf:"varint,15,opt,name=resCode,proto3" json:"resCode,omitempty"`
-	ErrMessage *wrapperspb.StringValue `protobuf:"bytes,16,opt,name=errMessage,proto3" json:"errMessage,omitempty"`
-	StackTrace *wrapperspb.StringValue `protobuf:"bytes,17,opt,name=stackTrace,proto3" json:"stackTrace,omitempty"`
-	StartedAt  string                  `protobuf:"bytes,18,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
-	FinishedAt string                  `protobuf:"bytes,19,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RequestHttpCallV1) Reset() {
 	*x = RequestHttpCallV1{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[3]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[3]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestHttpCallV1) String() string {
@@ -435,7 +449,7 @@ func (*RequestHttpCallV1) ProtoMessage() {}
 
 func (x *RequestHttpCallV1) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[3]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -585,10 +599,7 @@ func (x *RequestHttpCallV1) GetFinishedAt() string {
 
 // startedAt : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 type RequestServicePieceV1 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState  `protogen:"open.v1"`
 	Uid              string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
 	SvcName          string                  `protobuf:"bytes,2,opt,name=svcName,proto3" json:"svcName,omitempty"`
 	SvcVersion       string                  `protobuf:"bytes,3,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
@@ -605,15 +616,15 @@ type RequestServicePieceV1 struct {
 	ReqBody          *wrapperspb.StringValue `protobuf:"bytes,14,opt,name=reqBody,proto3" json:"reqBody,omitempty"`
 	ClientIp         string                  `protobuf:"bytes,15,opt,name=clientIp,proto3" json:"clientIp,omitempty"`
 	StartedAt        string                  `protobuf:"bytes,16,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *RequestServicePieceV1) Reset() {
 	*x = RequestServicePieceV1{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[4]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[4]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestServicePieceV1) String() string {
@@ -624,7 +635,7 @@ func (*RequestServicePieceV1) ProtoMessage() {}
 
 func (x *RequestServicePieceV1) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[4]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -755,10 +766,7 @@ func (x *RequestServicePieceV1) GetStartedAt() string {
 // startedAt  : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 // finishedAt : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
 type RequestServiceV1 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState  `protogen:"open.v1"`
 	Uid              string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
 	UserId           *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
 	PartnerId        *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
@@ -787,15 +795,15 @@ type RequestServiceV1 struct {
 	ClientIp         string                  `protobuf:"bytes,26,opt,name=clientIp,proto3" json:"clientIp,omitempty"`
 	StartedAt        string                  `protobuf:"bytes,27,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
 	FinishedAt       string                  `protobuf:"bytes,28,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *RequestServiceV1) Reset() {
 	*x = RequestServiceV1{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[5]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[5]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestServiceV1) String() string {
@@ -806,7 +814,7 @@ func (*RequestServiceV1) ProtoMessage() {}
 
 func (x *RequestServiceV1) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[5]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1019,10 +1027,7 @@ func (x *RequestServiceV1) GetFinishedAt() string {
 
 // severity : info, warning, error
 type RequestGrpcV1 struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
-	unknownFields protoimpl.UnknownFields
-
+	state            protoimpl.MessageState  `protogen:"open.v1"`
 	Uid              string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
 	UserId           *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
 	PartnerId        *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
@@ -1036,15 +1041,19 @@ type RequestGrpcV1 struct {
 	ExecFunction     string                  `protobuf:"bytes,11,opt,name=execFunction,proto3" json:"execFunction,omitempty"`
 	ReqHeader        *wrapperspb.StringValue `protobuf:"bytes,12,opt,name=reqHeader,proto3" json:"reqHeader,omitempty"`
 	Data             *wrapperspb.StringValue `protobuf:"bytes,13,opt,name=data,proto3" json:"data,omitempty"`
+	ErrMessage       *wrapperspb.StringValue `protobuf:"bytes,14,opt,name=errMessage,proto3" json:"errMessage,omitempty"`
+	StackTrace       *wrapperspb.StringValue `protobuf:"bytes,15,opt,name=stackTrace,proto3" json:"stackTrace,omitempty"`
+	StartedAt        string                  `protobuf:"bytes,16,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
+	FinishedAt       string                  `protobuf:"bytes,17,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
+	unknownFields    protoimpl.UnknownFields
+	sizeCache        protoimpl.SizeCache
 }
 
 func (x *RequestGrpcV1) Reset() {
 	*x = RequestGrpcV1{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_res_proto_clog_proto_msgTypes[6]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_res_proto_clog_proto_msgTypes[6]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RequestGrpcV1) String() string {
@@ -1055,7 +1064,7 @@ func (*RequestGrpcV1) ProtoMessage() {}
 
 func (x *RequestGrpcV1) ProtoReflect() protoreflect.Message {
 	mi := &file_res_proto_clog_proto_msgTypes[6]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -1161,346 +1170,400 @@ func (x *RequestGrpcV1) GetData() *wrapperspb.StringValue {
 	return nil
 }
 
+func (x *RequestGrpcV1) GetErrMessage() *wrapperspb.StringValue {
+	if x != nil {
+		return x.ErrMessage
+	}
+	return nil
+}
+
+func (x *RequestGrpcV1) GetStackTrace() *wrapperspb.StringValue {
+	if x != nil {
+		return x.StackTrace
+	}
+	return nil
+}
+
+func (x *RequestGrpcV1) GetStartedAt() string {
+	if x != nil {
+		return x.StartedAt
+	}
+	return ""
+}
+
+func (x *RequestGrpcV1) GetFinishedAt() string {
+	if x != nil {
+		return x.FinishedAt
+	}
+	return ""
+}
+
+// startedAt  : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
+// finishedAt : format "yyyy-MM-dd HH:mm:ss.SSSSSS TZ"
+type RequestDistLockV1 struct {
+	state         protoimpl.MessageState  `protogen:"open.v1"`
+	Uid           string                  `protobuf:"bytes,1,opt,name=uid,proto3" json:"uid,omitempty"`
+	UserId        *wrapperspb.StringValue `protobuf:"bytes,2,opt,name=userId,proto3" json:"userId,omitempty"`
+	PartnerId     *wrapperspb.StringValue `protobuf:"bytes,3,opt,name=partnerId,proto3" json:"partnerId,omitempty"`
+	SvcName       string                  `protobuf:"bytes,4,opt,name=svcName,proto3" json:"svcName,omitempty"`
+	SvcVersion    string                  `protobuf:"bytes,5,opt,name=svcVersion,proto3" json:"svcVersion,omitempty"`
+	Engine        string                  `protobuf:"bytes,6,opt,name=engine,proto3" json:"engine,omitempty"`
+	Address       string                  `protobuf:"bytes,7,opt,name=address,proto3" json:"address,omitempty"`
+	Key           string                  `protobuf:"bytes,8,opt,name=key,proto3" json:"key,omitempty"`
+	ErrWhen       *wrapperspb.StringValue `protobuf:"bytes,9,opt,name=errWhen,proto3" json:"errWhen,omitempty"`
+	ErrMessage    *wrapperspb.StringValue `protobuf:"bytes,10,opt,name=errMessage,proto3" json:"errMessage,omitempty"`
+	StackTrace    *wrapperspb.StringValue `protobuf:"bytes,11,opt,name=stackTrace,proto3" json:"stackTrace,omitempty"`
+	StartedAt     string                  `protobuf:"bytes,12,opt,name=startedAt,proto3" json:"startedAt,omitempty"`
+	FinishedAt    string                  `protobuf:"bytes,13,opt,name=finishedAt,proto3" json:"finishedAt,omitempty"`
+	unknownFields protoimpl.UnknownFields
+	sizeCache     protoimpl.SizeCache
+}
+
+func (x *RequestDistLockV1) Reset() {
+	*x = RequestDistLockV1{}
+	mi := &file_res_proto_clog_proto_msgTypes[7]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
+}
+
+func (x *RequestDistLockV1) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*RequestDistLockV1) ProtoMessage() {}
+
+func (x *RequestDistLockV1) ProtoReflect() protoreflect.Message {
+	mi := &file_res_proto_clog_proto_msgTypes[7]
+	if x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use RequestDistLockV1.ProtoReflect.Descriptor instead.
+func (*RequestDistLockV1) Descriptor() ([]byte, []int) {
+	return file_res_proto_clog_proto_rawDescGZIP(), []int{7}
+}
+
+func (x *RequestDistLockV1) GetUid() string {
+	if x != nil {
+		return x.Uid
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetUserId() *wrapperspb.StringValue {
+	if x != nil {
+		return x.UserId
+	}
+	return nil
+}
+
+func (x *RequestDistLockV1) GetPartnerId() *wrapperspb.StringValue {
+	if x != nil {
+		return x.PartnerId
+	}
+	return nil
+}
+
+func (x *RequestDistLockV1) GetSvcName() string {
+	if x != nil {
+		return x.SvcName
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetSvcVersion() string {
+	if x != nil {
+		return x.SvcVersion
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetEngine() string {
+	if x != nil {
+		return x.Engine
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetAddress() string {
+	if x != nil {
+		return x.Address
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetKey() string {
+	if x != nil {
+		return x.Key
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetErrWhen() *wrapperspb.StringValue {
+	if x != nil {
+		return x.ErrWhen
+	}
+	return nil
+}
+
+func (x *RequestDistLockV1) GetErrMessage() *wrapperspb.StringValue {
+	if x != nil {
+		return x.ErrMessage
+	}
+	return nil
+}
+
+func (x *RequestDistLockV1) GetStackTrace() *wrapperspb.StringValue {
+	if x != nil {
+		return x.StackTrace
+	}
+	return nil
+}
+
+func (x *RequestDistLockV1) GetStartedAt() string {
+	if x != nil {
+		return x.StartedAt
+	}
+	return ""
+}
+
+func (x *RequestDistLockV1) GetFinishedAt() string {
+	if x != nil {
+		return x.FinishedAt
+	}
+	return ""
+}
+
 var File_res_proto_clog_proto protoreflect.FileDescriptor
 
-var file_res_proto_clog_proto_rawDesc = []byte{
-	0x0a, 0x14, 0x72, 0x65, 0x73, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x2f, 0x63, 0x6c, 0x6f, 0x67,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x05, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x1a, 0x1e, 0x67,
-	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2f, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2f, 0x77,
-	0x72, 0x61, 0x70, 0x70, 0x65, 0x72, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x22, 0x3c, 0x0a,
-	0x08, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x16, 0x0a, 0x06, 0x73, 0x74, 0x61,
-	0x74, 0x75, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x06, 0x73, 0x74, 0x61, 0x74, 0x75,
-	0x73, 0x12, 0x18, 0x0a, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x18, 0x02, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x07, 0x6d, 0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x22, 0x87, 0x03, 0x0a, 0x0d,
-	0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x4e, 0x6f, 0x74, 0x65, 0x56, 0x31, 0x12, 0x10, 0x0a,
-	0x03, 0x75, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x69, 0x64, 0x12,
-	0x34, 0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x06, 0x75,
-	0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x3a, 0x0a, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72,
-	0x49, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e,
-	0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49,
-	0x64, 0x12, 0x18, 0x0a, 0x07, 0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x07, 0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x73,
-	0x76, 0x63, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x0a, 0x73, 0x76, 0x63, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x1a, 0x0a, 0x08, 0x65,
-	0x78, 0x65, 0x63, 0x50, 0x61, 0x74, 0x68, 0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x65,
-	0x78, 0x65, 0x63, 0x50, 0x61, 0x74, 0x68, 0x12, 0x22, 0x0a, 0x0c, 0x65, 0x78, 0x65, 0x63, 0x46,
-	0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x07, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x65,
-	0x78, 0x65, 0x63, 0x46, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x2e, 0x0a, 0x03, 0x6b,
-	0x65, 0x79, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e,
-	0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x03, 0x6b, 0x65, 0x79, 0x12, 0x34, 0x0a, 0x06, 0x73,
-	0x75, 0x62, 0x4b, 0x65, 0x79, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f,
-	0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74,
-	0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x06, 0x73, 0x75, 0x62, 0x4b, 0x65,
-	0x79, 0x12, 0x12, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x04, 0x64, 0x61, 0x74, 0x61, 0x22, 0xdd, 0x05, 0x0a, 0x0c, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x44, 0x62, 0x71, 0x56, 0x31, 0x12, 0x10, 0x0a, 0x03, 0x75, 0x69, 0x64, 0x18, 0x01, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x69, 0x64, 0x12, 0x34, 0x0a, 0x06, 0x75, 0x73, 0x65, 0x72,
-	0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e,
-	0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x12, 0x3a,
-	0x0a, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64, 0x18, 0x03, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52,
-	0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64, 0x12, 0x18, 0x0a, 0x07, 0x73, 0x76,
-	0x63, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x73, 0x76, 0x63,
-	0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x73, 0x76, 0x63, 0x56, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x73, 0x76, 0x63, 0x56, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x71, 0x6c, 0x51, 0x75, 0x65, 0x72, 0x79,
-	0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x71, 0x6c, 0x51, 0x75, 0x65, 0x72, 0x79,
-	0x12, 0x36, 0x0a, 0x07, 0x73, 0x71, 0x6c, 0x41, 0x72, 0x67, 0x73, 0x18, 0x07, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52,
-	0x07, 0x73, 0x71, 0x6c, 0x41, 0x72, 0x67, 0x73, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x65, 0x76, 0x65,
-	0x72, 0x69, 0x74, 0x79, 0x18, 0x08, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x65, 0x76, 0x65,
-	0x72, 0x69, 0x74, 0x79, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x78, 0x65, 0x63, 0x50, 0x61, 0x74, 0x68,
-	0x18, 0x09, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x78, 0x65, 0x63, 0x50, 0x61, 0x74, 0x68,
-	0x12, 0x22, 0x0a, 0x0c, 0x65, 0x78, 0x65, 0x63, 0x46, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e,
-	0x18, 0x0a, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x65, 0x78, 0x65, 0x63, 0x46, 0x75, 0x6e, 0x63,
-	0x74, 0x69, 0x6f, 0x6e, 0x12, 0x40, 0x0a, 0x0c, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x4d, 0x65, 0x73,
-	0x73, 0x61, 0x67, 0x65, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x0c, 0x65, 0x72, 0x72, 0x6f, 0x72, 0x4d,
-	0x65, 0x73, 0x73, 0x61, 0x67, 0x65, 0x12, 0x3c, 0x0a, 0x0a, 0x73, 0x74, 0x61, 0x63, 0x6b, 0x54,
-	0x72, 0x61, 0x63, 0x65, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x0a, 0x73, 0x74, 0x61, 0x63, 0x6b, 0x54,
-	0x72, 0x61, 0x63, 0x65, 0x12, 0x14, 0x0a, 0x05, 0x68, 0x6f, 0x73, 0x74, 0x31, 0x18, 0x0d, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x05, 0x68, 0x6f, 0x73, 0x74, 0x31, 0x12, 0x32, 0x0a, 0x05, 0x68, 0x6f,
-	0x73, 0x74, 0x32, 0x18, 0x0e, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67,
-	0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69,
-	0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x05, 0x68, 0x6f, 0x73, 0x74, 0x32, 0x12, 0x1c,
-	0x0a, 0x09, 0x64, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x31, 0x18, 0x0f, 0x20, 0x01, 0x28,
-	0x05, 0x52, 0x09, 0x64, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x31, 0x12, 0x39, 0x0a, 0x09,
-	0x64, 0x75, 0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x32, 0x18, 0x10, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1b, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x49, 0x6e, 0x74, 0x33, 0x32, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x64, 0x75,
-	0x72, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x32, 0x12, 0x1c, 0x0a, 0x09, 0x73, 0x74, 0x61, 0x72, 0x74,
-	0x65, 0x64, 0x41, 0x74, 0x18, 0x11, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x73, 0x74, 0x61, 0x72,
-	0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1e, 0x0a, 0x0a, 0x66, 0x69, 0x6e, 0x69, 0x73, 0x68, 0x65,
-	0x64, 0x41, 0x74, 0x18, 0x12, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x66, 0x69, 0x6e, 0x69, 0x73,
-	0x68, 0x65, 0x64, 0x41, 0x74, 0x22, 0xe5, 0x06, 0x0a, 0x11, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73,
-	0x74, 0x48, 0x74, 0x74, 0x70, 0x43, 0x61, 0x6c, 0x6c, 0x56, 0x31, 0x12, 0x10, 0x0a, 0x03, 0x75,
-	0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x69, 0x64, 0x12, 0x34, 0x0a,
-	0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x06, 0x75, 0x73, 0x65,
-	0x72, 0x49, 0x64, 0x12, 0x3a, 0x0a, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64,
-	0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56,
-	0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64, 0x12,
-	0x18, 0x0a, 0x07, 0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x07, 0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x73, 0x76, 0x63,
-	0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x73,
-	0x76, 0x63, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x10, 0x0a, 0x03, 0x75, 0x72, 0x6c,
-	0x18, 0x06, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x72, 0x6c, 0x12, 0x1a, 0x0a, 0x08, 0x73,
-	0x65, 0x76, 0x65, 0x72, 0x69, 0x74, 0x79, 0x18, 0x07, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73,
-	0x65, 0x76, 0x65, 0x72, 0x69, 0x74, 0x79, 0x12, 0x3a, 0x0a, 0x09, 0x72, 0x65, 0x71, 0x48, 0x65,
-	0x61, 0x64, 0x65, 0x72, 0x18, 0x08, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x72, 0x65, 0x71, 0x48, 0x65, 0x61,
-	0x64, 0x65, 0x72, 0x12, 0x38, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x18,
-	0x09, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61,
-	0x6c, 0x75, 0x65, 0x52, 0x08, 0x72, 0x65, 0x71, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x12, 0x38, 0x0a,
-	0x08, 0x72, 0x65, 0x71, 0x51, 0x75, 0x65, 0x72, 0x79, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x08, 0x72,
-	0x65, 0x71, 0x51, 0x75, 0x65, 0x72, 0x79, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x71, 0x46, 0x6f,
-	0x72, 0x6d, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e,
-	0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x07, 0x72, 0x65, 0x71, 0x46, 0x6f, 0x72, 0x6d, 0x12,
-	0x38, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x46, 0x69, 0x6c, 0x65, 0x73, 0x18, 0x0c, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52,
-	0x08, 0x72, 0x65, 0x71, 0x46, 0x69, 0x6c, 0x65, 0x73, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x71,
-	0x42, 0x6f, 0x64, 0x79, 0x18, 0x0d, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x07, 0x72, 0x65, 0x71, 0x42, 0x6f, 0x64,
-	0x79, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x73, 0x44, 0x61, 0x74, 0x61, 0x18, 0x0e, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65,
-	0x52, 0x07, 0x72, 0x65, 0x73, 0x44, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07, 0x72, 0x65, 0x73,
-	0x43, 0x6f, 0x64, 0x65, 0x18, 0x0f, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x72, 0x65, 0x73, 0x43,
-	0x6f, 0x64, 0x65, 0x12, 0x3c, 0x0a, 0x0a, 0x65, 0x72, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x18, 0x10, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67,
-	0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x0a, 0x65, 0x72, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61, 0x67,
-	0x65, 0x12, 0x3c, 0x0a, 0x0a, 0x73, 0x74, 0x61, 0x63, 0x6b, 0x54, 0x72, 0x61, 0x63, 0x65, 0x18,
-	0x11, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70,
-	0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61,
-	0x6c, 0x75, 0x65, 0x52, 0x0a, 0x73, 0x74, 0x61, 0x63, 0x6b, 0x54, 0x72, 0x61, 0x63, 0x65, 0x12,
-	0x1c, 0x0a, 0x09, 0x73, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x12, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x09, 0x73, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1e, 0x0a,
-	0x0a, 0x66, 0x69, 0x6e, 0x69, 0x73, 0x68, 0x65, 0x64, 0x41, 0x74, 0x18, 0x13, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x0a, 0x66, 0x69, 0x6e, 0x69, 0x73, 0x68, 0x65, 0x64, 0x41, 0x74, 0x22, 0xf3, 0x05,
-	0x0a, 0x15, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65,
-	0x50, 0x69, 0x65, 0x63, 0x65, 0x56, 0x31, 0x12, 0x10, 0x0a, 0x03, 0x75, 0x69, 0x64, 0x18, 0x01,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x69, 0x64, 0x12, 0x18, 0x0a, 0x07, 0x73, 0x76, 0x63,
-	0x4e, 0x61, 0x6d, 0x65, 0x18, 0x02, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x73, 0x76, 0x63, 0x4e,
-	0x61, 0x6d, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x73, 0x76, 0x63, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
-	0x6e, 0x18, 0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x73, 0x76, 0x63, 0x56, 0x65, 0x72, 0x73,
-	0x69, 0x6f, 0x6e, 0x12, 0x42, 0x0a, 0x0d, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74,
-	0x4e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x0d, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72,
-	0x65, 0x6e, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x48, 0x0a, 0x10, 0x73, 0x76, 0x63, 0x50, 0x61,
-	0x72, 0x65, 0x6e, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x05, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52,
-	0x10, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f,
-	0x6e, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x18, 0x06, 0x20,
-	0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74, 0x12, 0x10, 0x0a,
-	0x03, 0x75, 0x72, 0x6c, 0x18, 0x07, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x72, 0x6c, 0x12,
-	0x3c, 0x0a, 0x0a, 0x72, 0x65, 0x71, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x08, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75,
-	0x65, 0x52, 0x0a, 0x72, 0x65, 0x71, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x3a, 0x0a,
-	0x09, 0x72, 0x65, 0x71, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x18, 0x09, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09,
-	0x72, 0x65, 0x71, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x12, 0x3a, 0x0a, 0x09, 0x72, 0x65, 0x71,
-	0x48, 0x65, 0x61, 0x64, 0x65, 0x72, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67,
-	0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53,
-	0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x72, 0x65, 0x71, 0x48,
-	0x65, 0x61, 0x64, 0x65, 0x72, 0x12, 0x38, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x50, 0x61, 0x72, 0x61,
-	0x6d, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67,
-	0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x08, 0x72, 0x65, 0x71, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x12,
-	0x38, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x51, 0x75, 0x65, 0x72, 0x79, 0x18, 0x0c, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52,
-	0x08, 0x72, 0x65, 0x71, 0x51, 0x75, 0x65, 0x72, 0x79, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x71,
-	0x46, 0x6f, 0x72, 0x6d, 0x18, 0x0d, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x07, 0x72, 0x65, 0x71, 0x46, 0x6f, 0x72,
-	0x6d, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x71, 0x42, 0x6f, 0x64, 0x79, 0x18, 0x0e, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65,
-	0x52, 0x07, 0x72, 0x65, 0x71, 0x42, 0x6f, 0x64, 0x79, 0x12, 0x1a, 0x0a, 0x08, 0x63, 0x6c, 0x69,
-	0x65, 0x6e, 0x74, 0x49, 0x70, 0x18, 0x0f, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x63, 0x6c, 0x69,
-	0x65, 0x6e, 0x74, 0x49, 0x70, 0x12, 0x1c, 0x0a, 0x09, 0x73, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64,
-	0x41, 0x74, 0x18, 0x10, 0x20, 0x01, 0x28, 0x09, 0x52, 0x09, 0x73, 0x74, 0x61, 0x72, 0x74, 0x65,
-	0x64, 0x41, 0x74, 0x22, 0x84, 0x0a, 0x0a, 0x10, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x53,
-	0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x56, 0x31, 0x12, 0x10, 0x0a, 0x03, 0x75, 0x69, 0x64, 0x18,
-	0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x69, 0x64, 0x12, 0x34, 0x0a, 0x06, 0x75, 0x73,
-	0x65, 0x72, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f,
-	0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72,
-	0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64,
-	0x12, 0x3a, 0x0a, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64, 0x18, 0x03, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75,
-	0x65, 0x52, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64, 0x12, 0x18, 0x0a, 0x07,
-	0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x73,
-	0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x73, 0x76, 0x63, 0x56, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x73, 0x76, 0x63, 0x56,
-	0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x42, 0x0a, 0x0d, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72,
-	0x65, 0x6e, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e,
-	0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e,
-	0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x0d, 0x73, 0x76, 0x63,
-	0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x48, 0x0a, 0x10, 0x73, 0x76,
-	0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x07,
-	0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c,
-	0x75, 0x65, 0x52, 0x10, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x56, 0x65, 0x72,
-	0x73, 0x69, 0x6f, 0x6e, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74,
-	0x18, 0x08, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x6e, 0x64, 0x70, 0x6f, 0x69, 0x6e, 0x74,
-	0x12, 0x10, 0x0a, 0x03, 0x75, 0x72, 0x6c, 0x18, 0x09, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75,
-	0x72, 0x6c, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x65, 0x76, 0x65, 0x72, 0x69, 0x74, 0x79, 0x18, 0x0a,
-	0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x65, 0x76, 0x65, 0x72, 0x69, 0x74, 0x79, 0x12, 0x1a,
-	0x0a, 0x08, 0x65, 0x78, 0x65, 0x63, 0x50, 0x61, 0x74, 0x68, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x08, 0x65, 0x78, 0x65, 0x63, 0x50, 0x61, 0x74, 0x68, 0x12, 0x22, 0x0a, 0x0c, 0x65, 0x78,
-	0x65, 0x63, 0x46, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x09,
-	0x52, 0x0c, 0x65, 0x78, 0x65, 0x63, 0x46, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x3c,
-	0x0a, 0x0a, 0x72, 0x65, 0x71, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x0d, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65,
-	0x52, 0x0a, 0x72, 0x65, 0x71, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x3a, 0x0a, 0x09,
-	0x72, 0x65, 0x71, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x18, 0x0e, 0x20, 0x01, 0x28, 0x0b, 0x32,
-	0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75,
-	0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x72,
-	0x65, 0x71, 0x53, 0x6f, 0x75, 0x72, 0x63, 0x65, 0x12, 0x3a, 0x0a, 0x09, 0x72, 0x65, 0x71, 0x48,
-	0x65, 0x61, 0x64, 0x65, 0x72, 0x18, 0x0f, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f,
-	0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74,
-	0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x72, 0x65, 0x71, 0x48, 0x65,
-	0x61, 0x64, 0x65, 0x72, 0x12, 0x38, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x50, 0x61, 0x72, 0x61, 0x6d,
-	0x18, 0x10, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56,
-	0x61, 0x6c, 0x75, 0x65, 0x52, 0x08, 0x72, 0x65, 0x71, 0x50, 0x61, 0x72, 0x61, 0x6d, 0x12, 0x38,
-	0x0a, 0x08, 0x72, 0x65, 0x71, 0x51, 0x75, 0x65, 0x72, 0x79, 0x18, 0x11, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x08,
-	0x72, 0x65, 0x71, 0x51, 0x75, 0x65, 0x72, 0x79, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x71, 0x46,
-	0x6f, 0x72, 0x6d, 0x18, 0x12, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67,
-	0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69,
-	0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x07, 0x72, 0x65, 0x71, 0x46, 0x6f, 0x72, 0x6d,
-	0x12, 0x38, 0x0a, 0x08, 0x72, 0x65, 0x71, 0x46, 0x69, 0x6c, 0x65, 0x73, 0x18, 0x13, 0x20, 0x01,
-	0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74,
-	0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65,
-	0x52, 0x08, 0x72, 0x65, 0x71, 0x46, 0x69, 0x6c, 0x65, 0x73, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65,
-	0x71, 0x42, 0x6f, 0x64, 0x79, 0x18, 0x14, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f,
-	0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74,
-	0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x07, 0x72, 0x65, 0x71, 0x42, 0x6f,
-	0x64, 0x79, 0x12, 0x36, 0x0a, 0x07, 0x72, 0x65, 0x73, 0x44, 0x61, 0x74, 0x61, 0x18, 0x15, 0x20,
-	0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f,
-	0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75,
-	0x65, 0x52, 0x07, 0x72, 0x65, 0x73, 0x44, 0x61, 0x74, 0x61, 0x12, 0x18, 0x0a, 0x07, 0x72, 0x65,
-	0x73, 0x43, 0x6f, 0x64, 0x65, 0x18, 0x16, 0x20, 0x01, 0x28, 0x05, 0x52, 0x07, 0x72, 0x65, 0x73,
-	0x43, 0x6f, 0x64, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x72, 0x65, 0x73, 0x53, 0x75, 0x62, 0x43, 0x6f,
-	0x64, 0x65, 0x18, 0x17, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a, 0x72, 0x65, 0x73, 0x53, 0x75, 0x62,
-	0x43, 0x6f, 0x64, 0x65, 0x12, 0x3c, 0x0a, 0x0a, 0x65, 0x72, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x18, 0x18, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e,
-	0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x0a, 0x65, 0x72, 0x72, 0x4d, 0x65, 0x73, 0x73, 0x61,
-	0x67, 0x65, 0x12, 0x3c, 0x0a, 0x0a, 0x73, 0x74, 0x61, 0x63, 0x6b, 0x54, 0x72, 0x61, 0x63, 0x65,
-	0x18, 0x19, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e,
-	0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56,
-	0x61, 0x6c, 0x75, 0x65, 0x52, 0x0a, 0x73, 0x74, 0x61, 0x63, 0x6b, 0x54, 0x72, 0x61, 0x63, 0x65,
-	0x12, 0x1a, 0x0a, 0x08, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x49, 0x70, 0x18, 0x1a, 0x20, 0x01,
-	0x28, 0x09, 0x52, 0x08, 0x63, 0x6c, 0x69, 0x65, 0x6e, 0x74, 0x49, 0x70, 0x12, 0x1c, 0x0a, 0x09,
-	0x73, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64, 0x41, 0x74, 0x18, 0x1b, 0x20, 0x01, 0x28, 0x09, 0x52,
-	0x09, 0x73, 0x74, 0x61, 0x72, 0x74, 0x65, 0x64, 0x41, 0x74, 0x12, 0x1e, 0x0a, 0x0a, 0x66, 0x69,
-	0x6e, 0x69, 0x73, 0x68, 0x65, 0x64, 0x41, 0x74, 0x18, 0x1c, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a,
-	0x66, 0x69, 0x6e, 0x69, 0x73, 0x68, 0x65, 0x64, 0x41, 0x74, 0x22, 0xc7, 0x04, 0x0a, 0x0d, 0x52,
-	0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x47, 0x72, 0x70, 0x63, 0x56, 0x31, 0x12, 0x10, 0x0a, 0x03,
-	0x75, 0x69, 0x64, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x03, 0x75, 0x69, 0x64, 0x12, 0x34,
-	0x0a, 0x06, 0x75, 0x73, 0x65, 0x72, 0x49, 0x64, 0x18, 0x02, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c,
-	0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66,
-	0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x06, 0x75, 0x73,
-	0x65, 0x72, 0x49, 0x64, 0x12, 0x3a, 0x0a, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49,
-	0x64, 0x18, 0x03, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65,
-	0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67,
-	0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x70, 0x61, 0x72, 0x74, 0x6e, 0x65, 0x72, 0x49, 0x64,
-	0x12, 0x18, 0x0a, 0x07, 0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x04, 0x20, 0x01, 0x28,
-	0x09, 0x52, 0x07, 0x73, 0x76, 0x63, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x1e, 0x0a, 0x0a, 0x73, 0x76,
-	0x63, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x18, 0x05, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0a,
-	0x73, 0x76, 0x63, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x42, 0x0a, 0x0d, 0x73, 0x76,
-	0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x18, 0x06, 0x20, 0x01, 0x28,
-	0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f,
-	0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52,
-	0x0d, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x4e, 0x61, 0x6d, 0x65, 0x12, 0x48,
-	0x0a, 0x10, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72, 0x65, 0x6e, 0x74, 0x56, 0x65, 0x72, 0x73, 0x69,
-	0x6f, 0x6e, 0x18, 0x07, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c,
-	0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e,
-	0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x10, 0x73, 0x76, 0x63, 0x50, 0x61, 0x72, 0x65, 0x6e,
-	0x74, 0x56, 0x65, 0x72, 0x73, 0x69, 0x6f, 0x6e, 0x12, 0x20, 0x0a, 0x0b, 0x64, 0x65, 0x73, 0x74,
-	0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x18, 0x08, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0b, 0x64,
-	0x65, 0x73, 0x74, 0x69, 0x6e, 0x61, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x1a, 0x0a, 0x08, 0x73, 0x65,
-	0x76, 0x65, 0x72, 0x69, 0x74, 0x79, 0x18, 0x09, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x73, 0x65,
-	0x76, 0x65, 0x72, 0x69, 0x74, 0x79, 0x12, 0x1a, 0x0a, 0x08, 0x65, 0x78, 0x65, 0x63, 0x50, 0x61,
-	0x74, 0x68, 0x18, 0x0a, 0x20, 0x01, 0x28, 0x09, 0x52, 0x08, 0x65, 0x78, 0x65, 0x63, 0x50, 0x61,
-	0x74, 0x68, 0x12, 0x22, 0x0a, 0x0c, 0x65, 0x78, 0x65, 0x63, 0x46, 0x75, 0x6e, 0x63, 0x74, 0x69,
-	0x6f, 0x6e, 0x18, 0x0b, 0x20, 0x01, 0x28, 0x09, 0x52, 0x0c, 0x65, 0x78, 0x65, 0x63, 0x46, 0x75,
-	0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x12, 0x3a, 0x0a, 0x09, 0x72, 0x65, 0x71, 0x48, 0x65, 0x61,
-	0x64, 0x65, 0x72, 0x18, 0x0c, 0x20, 0x01, 0x28, 0x0b, 0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67,
-	0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62, 0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69,
-	0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x09, 0x72, 0x65, 0x71, 0x48, 0x65, 0x61, 0x64,
-	0x65, 0x72, 0x12, 0x30, 0x0a, 0x04, 0x64, 0x61, 0x74, 0x61, 0x18, 0x0d, 0x20, 0x01, 0x28, 0x0b,
-	0x32, 0x1c, 0x2e, 0x67, 0x6f, 0x6f, 0x67, 0x6c, 0x65, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x62,
-	0x75, 0x66, 0x2e, 0x53, 0x74, 0x72, 0x69, 0x6e, 0x67, 0x56, 0x61, 0x6c, 0x75, 0x65, 0x52, 0x04,
-	0x64, 0x61, 0x74, 0x61, 0x32, 0xcf, 0x02, 0x0a, 0x0b, 0x43, 0x4c, 0x6f, 0x67, 0x53, 0x65, 0x72,
-	0x76, 0x69, 0x63, 0x65, 0x12, 0x2f, 0x0a, 0x06, 0x4e, 0x6f, 0x74, 0x65, 0x56, 0x31, 0x12, 0x14,
-	0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x4e, 0x6f,
-	0x74, 0x65, 0x56, 0x31, 0x1a, 0x0f, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x2d, 0x0a, 0x05, 0x44, 0x62, 0x71, 0x56, 0x31, 0x12, 0x13,
-	0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x44, 0x62,
-	0x71, 0x56, 0x31, 0x1a, 0x0f, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x73, 0x70,
-	0x6f, 0x6e, 0x73, 0x65, 0x12, 0x37, 0x0a, 0x0a, 0x48, 0x74, 0x74, 0x70, 0x43, 0x61, 0x6c, 0x6c,
-	0x56, 0x31, 0x12, 0x18, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65,
-	0x73, 0x74, 0x48, 0x74, 0x74, 0x70, 0x43, 0x61, 0x6c, 0x6c, 0x56, 0x31, 0x1a, 0x0f, 0x2e, 0x73,
-	0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x3f, 0x0a,
-	0x0e, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x50, 0x69, 0x65, 0x63, 0x65, 0x56, 0x31, 0x12,
-	0x1c, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x53,
-	0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x50, 0x69, 0x65, 0x63, 0x65, 0x56, 0x31, 0x1a, 0x0f, 0x2e,
-	0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x35,
-	0x0a, 0x09, 0x53, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x56, 0x31, 0x12, 0x17, 0x2e, 0x73, 0x63,
-	0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x53, 0x65, 0x72, 0x76, 0x69,
-	0x63, 0x65, 0x56, 0x31, 0x1a, 0x0f, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x73,
-	0x70, 0x6f, 0x6e, 0x73, 0x65, 0x12, 0x2f, 0x0a, 0x06, 0x47, 0x72, 0x70, 0x63, 0x56, 0x31, 0x12,
-	0x14, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65, 0x71, 0x75, 0x65, 0x73, 0x74, 0x47,
-	0x72, 0x70, 0x63, 0x56, 0x31, 0x1a, 0x0f, 0x2e, 0x73, 0x63, 0x6c, 0x6f, 0x67, 0x2e, 0x52, 0x65,
-	0x73, 0x70, 0x6f, 0x6e, 0x73, 0x65, 0x42, 0x1b, 0x5a, 0x19, 0x72, 0x65, 0x73, 0x2f, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x2f, 0x67, 0x65, 0x6e, 0x65, 0x72, 0x61, 0x74, 0x65, 0x64, 0x2f, 0x73, 0x63,
-	0x6c, 0x6f, 0x67, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x33,
-}
+const file_res_proto_clog_proto_rawDesc = "" +
+	"\n" +
+	"\x14res/proto/clog.proto\x12\x05sclog\x1a\x1egoogle/protobuf/wrappers.proto\"<\n" +
+	"\bResponse\x12\x16\n" +
+	"\x06status\x18\x01 \x01(\tR\x06status\x12\x18\n" +
+	"\amessage\x18\x02 \x01(\tR\amessage\"\xa7\x03\n" +
+	"\rRequestNoteV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x124\n" +
+	"\x06userId\x18\x02 \x01(\v2\x1c.google.protobuf.StringValueR\x06userId\x12:\n" +
+	"\tpartnerId\x18\x03 \x01(\v2\x1c.google.protobuf.StringValueR\tpartnerId\x12\x18\n" +
+	"\asvcName\x18\x04 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x05 \x01(\tR\n" +
+	"svcVersion\x12\x1a\n" +
+	"\bexecPath\x18\x06 \x01(\tR\bexecPath\x12\"\n" +
+	"\fexecFunction\x18\a \x01(\tR\fexecFunction\x12.\n" +
+	"\x03key\x18\b \x01(\v2\x1c.google.protobuf.StringValueR\x03key\x124\n" +
+	"\x06subKey\x18\t \x01(\v2\x1c.google.protobuf.StringValueR\x06subKey\x12\x12\n" +
+	"\x04data\x18\n" +
+	" \x01(\tR\x04data\x12\x1e\n" +
+	"\n" +
+	"occurredAt\x18\v \x01(\tR\n" +
+	"occurredAt\"\xd1\x06\n" +
+	"\fRequestDbqV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x124\n" +
+	"\x06userId\x18\x02 \x01(\v2\x1c.google.protobuf.StringValueR\x06userId\x12:\n" +
+	"\tpartnerId\x18\x03 \x01(\v2\x1c.google.protobuf.StringValueR\tpartnerId\x12\x18\n" +
+	"\asvcName\x18\x04 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x05 \x01(\tR\n" +
+	"svcVersion\x12\x1a\n" +
+	"\bsqlQuery\x18\x06 \x01(\tR\bsqlQuery\x126\n" +
+	"\asqlArgs\x18\a \x01(\v2\x1c.google.protobuf.StringValueR\asqlArgs\x12\x1a\n" +
+	"\bseverity\x18\b \x01(\tR\bseverity\x12\x1a\n" +
+	"\bexecPath\x18\t \x01(\tR\bexecPath\x12\"\n" +
+	"\fexecFunction\x18\n" +
+	" \x01(\tR\fexecFunction\x12@\n" +
+	"\ferrorMessage\x18\v \x01(\v2\x1c.google.protobuf.StringValueR\ferrorMessage\x12<\n" +
+	"\n" +
+	"stackTrace\x18\f \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"stackTrace\x124\n" +
+	"\x06dbName\x18\r \x01(\v2\x1c.google.protobuf.StringValueR\x06dbName\x12<\n" +
+	"\n" +
+	"schemaName\x18\x0e \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"schemaName\x12\x14\n" +
+	"\x05host1\x18\x0f \x01(\tR\x05host1\x122\n" +
+	"\x05host2\x18\x10 \x01(\v2\x1c.google.protobuf.StringValueR\x05host2\x12\x1c\n" +
+	"\tduration1\x18\x11 \x01(\x05R\tduration1\x129\n" +
+	"\tduration2\x18\x12 \x01(\v2\x1b.google.protobuf.Int32ValueR\tduration2\x12\x1c\n" +
+	"\tstartedAt\x18\x13 \x01(\tR\tstartedAt\x12\x1e\n" +
+	"\n" +
+	"finishedAt\x18\x14 \x01(\tR\n" +
+	"finishedAt\"\xe5\x06\n" +
+	"\x11RequestHttpCallV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x124\n" +
+	"\x06userId\x18\x02 \x01(\v2\x1c.google.protobuf.StringValueR\x06userId\x12:\n" +
+	"\tpartnerId\x18\x03 \x01(\v2\x1c.google.protobuf.StringValueR\tpartnerId\x12\x18\n" +
+	"\asvcName\x18\x04 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x05 \x01(\tR\n" +
+	"svcVersion\x12\x10\n" +
+	"\x03url\x18\x06 \x01(\tR\x03url\x12\x1a\n" +
+	"\bseverity\x18\a \x01(\tR\bseverity\x12:\n" +
+	"\treqHeader\x18\b \x01(\v2\x1c.google.protobuf.StringValueR\treqHeader\x128\n" +
+	"\breqParam\x18\t \x01(\v2\x1c.google.protobuf.StringValueR\breqParam\x128\n" +
+	"\breqQuery\x18\n" +
+	" \x01(\v2\x1c.google.protobuf.StringValueR\breqQuery\x126\n" +
+	"\areqForm\x18\v \x01(\v2\x1c.google.protobuf.StringValueR\areqForm\x128\n" +
+	"\breqFiles\x18\f \x01(\v2\x1c.google.protobuf.StringValueR\breqFiles\x126\n" +
+	"\areqBody\x18\r \x01(\v2\x1c.google.protobuf.StringValueR\areqBody\x126\n" +
+	"\aresData\x18\x0e \x01(\v2\x1c.google.protobuf.StringValueR\aresData\x12\x18\n" +
+	"\aresCode\x18\x0f \x01(\x05R\aresCode\x12<\n" +
+	"\n" +
+	"errMessage\x18\x10 \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"errMessage\x12<\n" +
+	"\n" +
+	"stackTrace\x18\x11 \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"stackTrace\x12\x1c\n" +
+	"\tstartedAt\x18\x12 \x01(\tR\tstartedAt\x12\x1e\n" +
+	"\n" +
+	"finishedAt\x18\x13 \x01(\tR\n" +
+	"finishedAt\"\xf3\x05\n" +
+	"\x15RequestServicePieceV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x12\x18\n" +
+	"\asvcName\x18\x02 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x03 \x01(\tR\n" +
+	"svcVersion\x12B\n" +
+	"\rsvcParentName\x18\x04 \x01(\v2\x1c.google.protobuf.StringValueR\rsvcParentName\x12H\n" +
+	"\x10svcParentVersion\x18\x05 \x01(\v2\x1c.google.protobuf.StringValueR\x10svcParentVersion\x12\x1a\n" +
+	"\bendpoint\x18\x06 \x01(\tR\bendpoint\x12\x10\n" +
+	"\x03url\x18\a \x01(\tR\x03url\x12<\n" +
+	"\n" +
+	"reqVersion\x18\b \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"reqVersion\x12:\n" +
+	"\treqSource\x18\t \x01(\v2\x1c.google.protobuf.StringValueR\treqSource\x12:\n" +
+	"\treqHeader\x18\n" +
+	" \x01(\v2\x1c.google.protobuf.StringValueR\treqHeader\x128\n" +
+	"\breqParam\x18\v \x01(\v2\x1c.google.protobuf.StringValueR\breqParam\x128\n" +
+	"\breqQuery\x18\f \x01(\v2\x1c.google.protobuf.StringValueR\breqQuery\x126\n" +
+	"\areqForm\x18\r \x01(\v2\x1c.google.protobuf.StringValueR\areqForm\x126\n" +
+	"\areqBody\x18\x0e \x01(\v2\x1c.google.protobuf.StringValueR\areqBody\x12\x1a\n" +
+	"\bclientIp\x18\x0f \x01(\tR\bclientIp\x12\x1c\n" +
+	"\tstartedAt\x18\x10 \x01(\tR\tstartedAt\"\x84\n" +
+	"\n" +
+	"\x10RequestServiceV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x124\n" +
+	"\x06userId\x18\x02 \x01(\v2\x1c.google.protobuf.StringValueR\x06userId\x12:\n" +
+	"\tpartnerId\x18\x03 \x01(\v2\x1c.google.protobuf.StringValueR\tpartnerId\x12\x18\n" +
+	"\asvcName\x18\x04 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x05 \x01(\tR\n" +
+	"svcVersion\x12B\n" +
+	"\rsvcParentName\x18\x06 \x01(\v2\x1c.google.protobuf.StringValueR\rsvcParentName\x12H\n" +
+	"\x10svcParentVersion\x18\a \x01(\v2\x1c.google.protobuf.StringValueR\x10svcParentVersion\x12\x1a\n" +
+	"\bendpoint\x18\b \x01(\tR\bendpoint\x12\x10\n" +
+	"\x03url\x18\t \x01(\tR\x03url\x12\x1a\n" +
+	"\bseverity\x18\n" +
+	" \x01(\tR\bseverity\x12\x1a\n" +
+	"\bexecPath\x18\v \x01(\tR\bexecPath\x12\"\n" +
+	"\fexecFunction\x18\f \x01(\tR\fexecFunction\x12<\n" +
+	"\n" +
+	"reqVersion\x18\r \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"reqVersion\x12:\n" +
+	"\treqSource\x18\x0e \x01(\v2\x1c.google.protobuf.StringValueR\treqSource\x12:\n" +
+	"\treqHeader\x18\x0f \x01(\v2\x1c.google.protobuf.StringValueR\treqHeader\x128\n" +
+	"\breqParam\x18\x10 \x01(\v2\x1c.google.protobuf.StringValueR\breqParam\x128\n" +
+	"\breqQuery\x18\x11 \x01(\v2\x1c.google.protobuf.StringValueR\breqQuery\x126\n" +
+	"\areqForm\x18\x12 \x01(\v2\x1c.google.protobuf.StringValueR\areqForm\x128\n" +
+	"\breqFiles\x18\x13 \x01(\v2\x1c.google.protobuf.StringValueR\breqFiles\x126\n" +
+	"\areqBody\x18\x14 \x01(\v2\x1c.google.protobuf.StringValueR\areqBody\x126\n" +
+	"\aresData\x18\x15 \x01(\v2\x1c.google.protobuf.StringValueR\aresData\x12\x18\n" +
+	"\aresCode\x18\x16 \x01(\x05R\aresCode\x12\x1e\n" +
+	"\n" +
+	"resSubCode\x18\x17 \x01(\tR\n" +
+	"resSubCode\x12<\n" +
+	"\n" +
+	"errMessage\x18\x18 \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"errMessage\x12<\n" +
+	"\n" +
+	"stackTrace\x18\x19 \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"stackTrace\x12\x1a\n" +
+	"\bclientIp\x18\x1a \x01(\tR\bclientIp\x12\x1c\n" +
+	"\tstartedAt\x18\x1b \x01(\tR\tstartedAt\x12\x1e\n" +
+	"\n" +
+	"finishedAt\x18\x1c \x01(\tR\n" +
+	"finishedAt\"\x81\x06\n" +
+	"\rRequestGrpcV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x124\n" +
+	"\x06userId\x18\x02 \x01(\v2\x1c.google.protobuf.StringValueR\x06userId\x12:\n" +
+	"\tpartnerId\x18\x03 \x01(\v2\x1c.google.protobuf.StringValueR\tpartnerId\x12\x18\n" +
+	"\asvcName\x18\x04 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x05 \x01(\tR\n" +
+	"svcVersion\x12B\n" +
+	"\rsvcParentName\x18\x06 \x01(\v2\x1c.google.protobuf.StringValueR\rsvcParentName\x12H\n" +
+	"\x10svcParentVersion\x18\a \x01(\v2\x1c.google.protobuf.StringValueR\x10svcParentVersion\x12 \n" +
+	"\vdestination\x18\b \x01(\tR\vdestination\x12\x1a\n" +
+	"\bseverity\x18\t \x01(\tR\bseverity\x12\x1a\n" +
+	"\bexecPath\x18\n" +
+	" \x01(\tR\bexecPath\x12\"\n" +
+	"\fexecFunction\x18\v \x01(\tR\fexecFunction\x12:\n" +
+	"\treqHeader\x18\f \x01(\v2\x1c.google.protobuf.StringValueR\treqHeader\x120\n" +
+	"\x04data\x18\r \x01(\v2\x1c.google.protobuf.StringValueR\x04data\x12<\n" +
+	"\n" +
+	"errMessage\x18\x0e \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"errMessage\x12<\n" +
+	"\n" +
+	"stackTrace\x18\x0f \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"stackTrace\x12\x1c\n" +
+	"\tstartedAt\x18\x10 \x01(\tR\tstartedAt\x12\x1e\n" +
+	"\n" +
+	"finishedAt\x18\x11 \x01(\tR\n" +
+	"finishedAt\"\x87\x04\n" +
+	"\x11RequestDistLockV1\x12\x10\n" +
+	"\x03uid\x18\x01 \x01(\tR\x03uid\x124\n" +
+	"\x06userId\x18\x02 \x01(\v2\x1c.google.protobuf.StringValueR\x06userId\x12:\n" +
+	"\tpartnerId\x18\x03 \x01(\v2\x1c.google.protobuf.StringValueR\tpartnerId\x12\x18\n" +
+	"\asvcName\x18\x04 \x01(\tR\asvcName\x12\x1e\n" +
+	"\n" +
+	"svcVersion\x18\x05 \x01(\tR\n" +
+	"svcVersion\x12\x16\n" +
+	"\x06engine\x18\x06 \x01(\tR\x06engine\x12\x18\n" +
+	"\aaddress\x18\a \x01(\tR\aaddress\x12\x10\n" +
+	"\x03key\x18\b \x01(\tR\x03key\x126\n" +
+	"\aerrWhen\x18\t \x01(\v2\x1c.google.protobuf.StringValueR\aerrWhen\x12<\n" +
+	"\n" +
+	"errMessage\x18\n" +
+	" \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"errMessage\x12<\n" +
+	"\n" +
+	"stackTrace\x18\v \x01(\v2\x1c.google.protobuf.StringValueR\n" +
+	"stackTrace\x12\x1c\n" +
+	"\tstartedAt\x18\f \x01(\tR\tstartedAt\x12\x1e\n" +
+	"\n" +
+	"finishedAt\x18\r \x01(\tR\n" +
+	"finishedAt2\x88\x03\n" +
+	"\vCLogService\x12/\n" +
+	"\x06NoteV1\x12\x14.sclog.RequestNoteV1\x1a\x0f.sclog.Response\x12-\n" +
+	"\x05DbqV1\x12\x13.sclog.RequestDbqV1\x1a\x0f.sclog.Response\x127\n" +
+	"\n" +
+	"HttpCallV1\x12\x18.sclog.RequestHttpCallV1\x1a\x0f.sclog.Response\x12?\n" +
+	"\x0eServicePieceV1\x12\x1c.sclog.RequestServicePieceV1\x1a\x0f.sclog.Response\x125\n" +
+	"\tServiceV1\x12\x17.sclog.RequestServiceV1\x1a\x0f.sclog.Response\x12/\n" +
+	"\x06GrpcV1\x12\x14.sclog.RequestGrpcV1\x1a\x0f.sclog.Response\x127\n" +
+	"\n" +
+	"DistLockV1\x12\x18.sclog.RequestDistLockV1\x1a\x0f.sclog.ResponseB\x1bZ\x19res/proto/generated/sclogb\x06proto3"
 
 var (
 	file_res_proto_clog_proto_rawDescOnce sync.Once
-	file_res_proto_clog_proto_rawDescData = file_res_proto_clog_proto_rawDesc
+	file_res_proto_clog_proto_rawDescData []byte
 )
 
 func file_res_proto_clog_proto_rawDescGZIP() []byte {
 	file_res_proto_clog_proto_rawDescOnce.Do(func() {
-		file_res_proto_clog_proto_rawDescData = protoimpl.X.CompressGZIP(file_res_proto_clog_proto_rawDescData)
+		file_res_proto_clog_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_res_proto_clog_proto_rawDesc), len(file_res_proto_clog_proto_rawDesc)))
 	})
 	return file_res_proto_clog_proto_rawDescData
 }
 
-var file_res_proto_clog_proto_msgTypes = make([]protoimpl.MessageInfo, 7)
-var file_res_proto_clog_proto_goTypes = []interface{}{
+var file_res_proto_clog_proto_msgTypes = make([]protoimpl.MessageInfo, 8)
+var file_res_proto_clog_proto_goTypes = []any{
 	(*Response)(nil),               // 0: sclog.Response
 	(*RequestNoteV1)(nil),          // 1: sclog.RequestNoteV1
 	(*RequestDbqV1)(nil),           // 2: sclog.RequestDbqV1
@@ -1508,79 +1571,91 @@ var file_res_proto_clog_proto_goTypes = []interface{}{
 	(*RequestServicePieceV1)(nil),  // 4: sclog.RequestServicePieceV1
 	(*RequestServiceV1)(nil),       // 5: sclog.RequestServiceV1
 	(*RequestGrpcV1)(nil),          // 6: sclog.RequestGrpcV1
-	(*wrapperspb.StringValue)(nil), // 7: google.protobuf.StringValue
-	(*wrapperspb.Int32Value)(nil),  // 8: google.protobuf.Int32Value
+	(*RequestDistLockV1)(nil),      // 7: sclog.RequestDistLockV1
+	(*wrapperspb.StringValue)(nil), // 8: google.protobuf.StringValue
+	(*wrapperspb.Int32Value)(nil),  // 9: google.protobuf.Int32Value
 }
 var file_res_proto_clog_proto_depIdxs = []int32{
-	7,  // 0: sclog.RequestNoteV1.userId:type_name -> google.protobuf.StringValue
-	7,  // 1: sclog.RequestNoteV1.partnerId:type_name -> google.protobuf.StringValue
-	7,  // 2: sclog.RequestNoteV1.key:type_name -> google.protobuf.StringValue
-	7,  // 3: sclog.RequestNoteV1.subKey:type_name -> google.protobuf.StringValue
-	7,  // 4: sclog.RequestDbqV1.userId:type_name -> google.protobuf.StringValue
-	7,  // 5: sclog.RequestDbqV1.partnerId:type_name -> google.protobuf.StringValue
-	7,  // 6: sclog.RequestDbqV1.sqlArgs:type_name -> google.protobuf.StringValue
-	7,  // 7: sclog.RequestDbqV1.errorMessage:type_name -> google.protobuf.StringValue
-	7,  // 8: sclog.RequestDbqV1.stackTrace:type_name -> google.protobuf.StringValue
-	7,  // 9: sclog.RequestDbqV1.host2:type_name -> google.protobuf.StringValue
-	8,  // 10: sclog.RequestDbqV1.duration2:type_name -> google.protobuf.Int32Value
-	7,  // 11: sclog.RequestHttpCallV1.userId:type_name -> google.protobuf.StringValue
-	7,  // 12: sclog.RequestHttpCallV1.partnerId:type_name -> google.protobuf.StringValue
-	7,  // 13: sclog.RequestHttpCallV1.reqHeader:type_name -> google.protobuf.StringValue
-	7,  // 14: sclog.RequestHttpCallV1.reqParam:type_name -> google.protobuf.StringValue
-	7,  // 15: sclog.RequestHttpCallV1.reqQuery:type_name -> google.protobuf.StringValue
-	7,  // 16: sclog.RequestHttpCallV1.reqForm:type_name -> google.protobuf.StringValue
-	7,  // 17: sclog.RequestHttpCallV1.reqFiles:type_name -> google.protobuf.StringValue
-	7,  // 18: sclog.RequestHttpCallV1.reqBody:type_name -> google.protobuf.StringValue
-	7,  // 19: sclog.RequestHttpCallV1.resData:type_name -> google.protobuf.StringValue
-	7,  // 20: sclog.RequestHttpCallV1.errMessage:type_name -> google.protobuf.StringValue
-	7,  // 21: sclog.RequestHttpCallV1.stackTrace:type_name -> google.protobuf.StringValue
-	7,  // 22: sclog.RequestServicePieceV1.svcParentName:type_name -> google.protobuf.StringValue
-	7,  // 23: sclog.RequestServicePieceV1.svcParentVersion:type_name -> google.protobuf.StringValue
-	7,  // 24: sclog.RequestServicePieceV1.reqVersion:type_name -> google.protobuf.StringValue
-	7,  // 25: sclog.RequestServicePieceV1.reqSource:type_name -> google.protobuf.StringValue
-	7,  // 26: sclog.RequestServicePieceV1.reqHeader:type_name -> google.protobuf.StringValue
-	7,  // 27: sclog.RequestServicePieceV1.reqParam:type_name -> google.protobuf.StringValue
-	7,  // 28: sclog.RequestServicePieceV1.reqQuery:type_name -> google.protobuf.StringValue
-	7,  // 29: sclog.RequestServicePieceV1.reqForm:type_name -> google.protobuf.StringValue
-	7,  // 30: sclog.RequestServicePieceV1.reqBody:type_name -> google.protobuf.StringValue
-	7,  // 31: sclog.RequestServiceV1.userId:type_name -> google.protobuf.StringValue
-	7,  // 32: sclog.RequestServiceV1.partnerId:type_name -> google.protobuf.StringValue
-	7,  // 33: sclog.RequestServiceV1.svcParentName:type_name -> google.protobuf.StringValue
-	7,  // 34: sclog.RequestServiceV1.svcParentVersion:type_name -> google.protobuf.StringValue
-	7,  // 35: sclog.RequestServiceV1.reqVersion:type_name -> google.protobuf.StringValue
-	7,  // 36: sclog.RequestServiceV1.reqSource:type_name -> google.protobuf.StringValue
-	7,  // 37: sclog.RequestServiceV1.reqHeader:type_name -> google.protobuf.StringValue
-	7,  // 38: sclog.RequestServiceV1.reqParam:type_name -> google.protobuf.StringValue
-	7,  // 39: sclog.RequestServiceV1.reqQuery:type_name -> google.protobuf.StringValue
-	7,  // 40: sclog.RequestServiceV1.reqForm:type_name -> google.protobuf.StringValue
-	7,  // 41: sclog.RequestServiceV1.reqFiles:type_name -> google.protobuf.StringValue
-	7,  // 42: sclog.RequestServiceV1.reqBody:type_name -> google.protobuf.StringValue
-	7,  // 43: sclog.RequestServiceV1.resData:type_name -> google.protobuf.StringValue
-	7,  // 44: sclog.RequestServiceV1.errMessage:type_name -> google.protobuf.StringValue
-	7,  // 45: sclog.RequestServiceV1.stackTrace:type_name -> google.protobuf.StringValue
-	7,  // 46: sclog.RequestGrpcV1.userId:type_name -> google.protobuf.StringValue
-	7,  // 47: sclog.RequestGrpcV1.partnerId:type_name -> google.protobuf.StringValue
-	7,  // 48: sclog.RequestGrpcV1.svcParentName:type_name -> google.protobuf.StringValue
-	7,  // 49: sclog.RequestGrpcV1.svcParentVersion:type_name -> google.protobuf.StringValue
-	7,  // 50: sclog.RequestGrpcV1.reqHeader:type_name -> google.protobuf.StringValue
-	7,  // 51: sclog.RequestGrpcV1.data:type_name -> google.protobuf.StringValue
-	1,  // 52: sclog.CLogService.NoteV1:input_type -> sclog.RequestNoteV1
-	2,  // 53: sclog.CLogService.DbqV1:input_type -> sclog.RequestDbqV1
-	3,  // 54: sclog.CLogService.HttpCallV1:input_type -> sclog.RequestHttpCallV1
-	4,  // 55: sclog.CLogService.ServicePieceV1:input_type -> sclog.RequestServicePieceV1
-	5,  // 56: sclog.CLogService.ServiceV1:input_type -> sclog.RequestServiceV1
-	6,  // 57: sclog.CLogService.GrpcV1:input_type -> sclog.RequestGrpcV1
-	0,  // 58: sclog.CLogService.NoteV1:output_type -> sclog.Response
-	0,  // 59: sclog.CLogService.DbqV1:output_type -> sclog.Response
-	0,  // 60: sclog.CLogService.HttpCallV1:output_type -> sclog.Response
-	0,  // 61: sclog.CLogService.ServicePieceV1:output_type -> sclog.Response
-	0,  // 62: sclog.CLogService.ServiceV1:output_type -> sclog.Response
-	0,  // 63: sclog.CLogService.GrpcV1:output_type -> sclog.Response
-	58, // [58:64] is the sub-list for method output_type
-	52, // [52:58] is the sub-list for method input_type
-	52, // [52:52] is the sub-list for extension type_name
-	52, // [52:52] is the sub-list for extension extendee
-	0,  // [0:52] is the sub-list for field type_name
+	8,  // 0: sclog.RequestNoteV1.userId:type_name -> google.protobuf.StringValue
+	8,  // 1: sclog.RequestNoteV1.partnerId:type_name -> google.protobuf.StringValue
+	8,  // 2: sclog.RequestNoteV1.key:type_name -> google.protobuf.StringValue
+	8,  // 3: sclog.RequestNoteV1.subKey:type_name -> google.protobuf.StringValue
+	8,  // 4: sclog.RequestDbqV1.userId:type_name -> google.protobuf.StringValue
+	8,  // 5: sclog.RequestDbqV1.partnerId:type_name -> google.protobuf.StringValue
+	8,  // 6: sclog.RequestDbqV1.sqlArgs:type_name -> google.protobuf.StringValue
+	8,  // 7: sclog.RequestDbqV1.errorMessage:type_name -> google.protobuf.StringValue
+	8,  // 8: sclog.RequestDbqV1.stackTrace:type_name -> google.protobuf.StringValue
+	8,  // 9: sclog.RequestDbqV1.dbName:type_name -> google.protobuf.StringValue
+	8,  // 10: sclog.RequestDbqV1.schemaName:type_name -> google.protobuf.StringValue
+	8,  // 11: sclog.RequestDbqV1.host2:type_name -> google.protobuf.StringValue
+	9,  // 12: sclog.RequestDbqV1.duration2:type_name -> google.protobuf.Int32Value
+	8,  // 13: sclog.RequestHttpCallV1.userId:type_name -> google.protobuf.StringValue
+	8,  // 14: sclog.RequestHttpCallV1.partnerId:type_name -> google.protobuf.StringValue
+	8,  // 15: sclog.RequestHttpCallV1.reqHeader:type_name -> google.protobuf.StringValue
+	8,  // 16: sclog.RequestHttpCallV1.reqParam:type_name -> google.protobuf.StringValue
+	8,  // 17: sclog.RequestHttpCallV1.reqQuery:type_name -> google.protobuf.StringValue
+	8,  // 18: sclog.RequestHttpCallV1.reqForm:type_name -> google.protobuf.StringValue
+	8,  // 19: sclog.RequestHttpCallV1.reqFiles:type_name -> google.protobuf.StringValue
+	8,  // 20: sclog.RequestHttpCallV1.reqBody:type_name -> google.protobuf.StringValue
+	8,  // 21: sclog.RequestHttpCallV1.resData:type_name -> google.protobuf.StringValue
+	8,  // 22: sclog.RequestHttpCallV1.errMessage:type_name -> google.protobuf.StringValue
+	8,  // 23: sclog.RequestHttpCallV1.stackTrace:type_name -> google.protobuf.StringValue
+	8,  // 24: sclog.RequestServicePieceV1.svcParentName:type_name -> google.protobuf.StringValue
+	8,  // 25: sclog.RequestServicePieceV1.svcParentVersion:type_name -> google.protobuf.StringValue
+	8,  // 26: sclog.RequestServicePieceV1.reqVersion:type_name -> google.protobuf.StringValue
+	8,  // 27: sclog.RequestServicePieceV1.reqSource:type_name -> google.protobuf.StringValue
+	8,  // 28: sclog.RequestServicePieceV1.reqHeader:type_name -> google.protobuf.StringValue
+	8,  // 29: sclog.RequestServicePieceV1.reqParam:type_name -> google.protobuf.StringValue
+	8,  // 30: sclog.RequestServicePieceV1.reqQuery:type_name -> google.protobuf.StringValue
+	8,  // 31: sclog.RequestServicePieceV1.reqForm:type_name -> google.protobuf.StringValue
+	8,  // 32: sclog.RequestServicePieceV1.reqBody:type_name -> google.protobuf.StringValue
+	8,  // 33: sclog.RequestServiceV1.userId:type_name -> google.protobuf.StringValue
+	8,  // 34: sclog.RequestServiceV1.partnerId:type_name -> google.protobuf.StringValue
+	8,  // 35: sclog.RequestServiceV1.svcParentName:type_name -> google.protobuf.StringValue
+	8,  // 36: sclog.RequestServiceV1.svcParentVersion:type_name -> google.protobuf.StringValue
+	8,  // 37: sclog.RequestServiceV1.reqVersion:type_name -> google.protobuf.StringValue
+	8,  // 38: sclog.RequestServiceV1.reqSource:type_name -> google.protobuf.StringValue
+	8,  // 39: sclog.RequestServiceV1.reqHeader:type_name -> google.protobuf.StringValue
+	8,  // 40: sclog.RequestServiceV1.reqParam:type_name -> google.protobuf.StringValue
+	8,  // 41: sclog.RequestServiceV1.reqQuery:type_name -> google.protobuf.StringValue
+	8,  // 42: sclog.RequestServiceV1.reqForm:type_name -> google.protobuf.StringValue
+	8,  // 43: sclog.RequestServiceV1.reqFiles:type_name -> google.protobuf.StringValue
+	8,  // 44: sclog.RequestServiceV1.reqBody:type_name -> google.protobuf.StringValue
+	8,  // 45: sclog.RequestServiceV1.resData:type_name -> google.protobuf.StringValue
+	8,  // 46: sclog.RequestServiceV1.errMessage:type_name -> google.protobuf.StringValue
+	8,  // 47: sclog.RequestServiceV1.stackTrace:type_name -> google.protobuf.StringValue
+	8,  // 48: sclog.RequestGrpcV1.userId:type_name -> google.protobuf.StringValue
+	8,  // 49: sclog.RequestGrpcV1.partnerId:type_name -> google.protobuf.StringValue
+	8,  // 50: sclog.RequestGrpcV1.svcParentName:type_name -> google.protobuf.StringValue
+	8,  // 51: sclog.RequestGrpcV1.svcParentVersion:type_name -> google.protobuf.StringValue
+	8,  // 52: sclog.RequestGrpcV1.reqHeader:type_name -> google.protobuf.StringValue
+	8,  // 53: sclog.RequestGrpcV1.data:type_name -> google.protobuf.StringValue
+	8,  // 54: sclog.RequestGrpcV1.errMessage:type_name -> google.protobuf.StringValue
+	8,  // 55: sclog.RequestGrpcV1.stackTrace:type_name -> google.protobuf.StringValue
+	8,  // 56: sclog.RequestDistLockV1.userId:type_name -> google.protobuf.StringValue
+	8,  // 57: sclog.RequestDistLockV1.partnerId:type_name -> google.protobuf.StringValue
+	8,  // 58: sclog.RequestDistLockV1.errWhen:type_name -> google.protobuf.StringValue
+	8,  // 59: sclog.RequestDistLockV1.errMessage:type_name -> google.protobuf.StringValue
+	8,  // 60: sclog.RequestDistLockV1.stackTrace:type_name -> google.protobuf.StringValue
+	1,  // 61: sclog.CLogService.NoteV1:input_type -> sclog.RequestNoteV1
+	2,  // 62: sclog.CLogService.DbqV1:input_type -> sclog.RequestDbqV1
+	3,  // 63: sclog.CLogService.HttpCallV1:input_type -> sclog.RequestHttpCallV1
+	4,  // 64: sclog.CLogService.ServicePieceV1:input_type -> sclog.RequestServicePieceV1
+	5,  // 65: sclog.CLogService.ServiceV1:input_type -> sclog.RequestServiceV1
+	6,  // 66: sclog.CLogService.GrpcV1:input_type -> sclog.RequestGrpcV1
+	7,  // 67: sclog.CLogService.DistLockV1:input_type -> sclog.RequestDistLockV1
+	0,  // 68: sclog.CLogService.NoteV1:output_type -> sclog.Response
+	0,  // 69: sclog.CLogService.DbqV1:output_type -> sclog.Response
+	0,  // 70: sclog.CLogService.HttpCallV1:output_type -> sclog.Response
+	0,  // 71: sclog.CLogService.ServicePieceV1:output_type -> sclog.Response
+	0,  // 72: sclog.CLogService.ServiceV1:output_type -> sclog.Response
+	0,  // 73: sclog.CLogService.GrpcV1:output_type -> sclog.Response
+	0,  // 74: sclog.CLogService.DistLockV1:output_type -> sclog.Response
+	68, // [68:75] is the sub-list for method output_type
+	61, // [61:68] is the sub-list for method input_type
+	61, // [61:61] is the sub-list for extension type_name
+	61, // [61:61] is the sub-list for extension extendee
+	0,  // [0:61] is the sub-list for field type_name
 }
 
 func init() { file_res_proto_clog_proto_init() }
@@ -1588,99 +1663,13 @@ func file_res_proto_clog_proto_init() {
 	if File_res_proto_clog_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_res_proto_clog_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*Response); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_res_proto_clog_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestNoteV1); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_res_proto_clog_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestDbqV1); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_res_proto_clog_proto_msgTypes[3].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestHttpCallV1); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_res_proto_clog_proto_msgTypes[4].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestServicePieceV1); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_res_proto_clog_proto_msgTypes[5].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestServiceV1); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_res_proto_clog_proto_msgTypes[6].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RequestGrpcV1); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_res_proto_clog_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_res_proto_clog_proto_rawDesc), len(file_res_proto_clog_proto_rawDesc)),
 			NumEnums:      0,
-			NumMessages:   7,
+			NumMessages:   8,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
@@ -1689,7 +1678,6 @@ func file_res_proto_clog_proto_init() {
 		MessageInfos:      file_res_proto_clog_proto_msgTypes,
 	}.Build()
 	File_res_proto_clog_proto = out.File
-	file_res_proto_clog_proto_rawDesc = nil
 	file_res_proto_clog_proto_goTypes = nil
 	file_res_proto_clog_proto_depIdxs = nil
 }
