@@ -12,12 +12,10 @@ package app
 import (
 	"github.com/ClickHouse/clickhouse-go/v2/lib/driver"
 	"github.com/andypangaribuan/gmod/ice"
-	qdb "github.com/questdb/go-questdb-client/v3"
 )
 
 var (
-	Db      ice.DbInstance
-	QdbPool *qdb.LineSenderPool
-	ChDb    driver.Conn
-	Env     *stuEnv
+	Db   ice.DbInstance
+	ChDb driver.Conn
+	Env  *stuEnv
 )
