@@ -12,7 +12,6 @@ package clog
 import (
 	"clog/db/entity"
 	"clog/db/repo"
-	"time"
 
 	"github.com/andypangaribuan/gmod/fm"
 	"github.com/andypangaribuan/gmod/gm"
@@ -21,20 +20,18 @@ import (
 
 func (stu *stuClog) httpCallV1(req *sclog.RequestHttpCallV1, _ map[string]string) (*sclog.Response, error) {
 	var (
-		startedAt  *time.Time
-		finishedAt *time.Time
-		duration   int
-		err        error
+		timenow  = gm.Util.Timenow()
+		duration int
 	)
 
-	startedAt, _ = gm.Conv.Time.ToTimeFull(req.StartedAt)
-	finishedAt, _ = gm.Conv.Time.ToTimeFull(req.FinishedAt)
+	startedAt, _ := gm.Conv.Time.ToTimeFull(req.StartedAt)
+	finishedAt, _ := gm.Conv.Time.ToTimeFull(req.FinishedAt)
 	if startedAt != nil && finishedAt != nil {
 		duration = int(finishedAt.Sub(*startedAt).Milliseconds())
 	}
 
 	e := &entity.HttpCallV1{
-		CreatedAt:    gm.Util.Timenow(),
+		CreatedAt:    timenow,
 		Uid:          req.Uid,
 		UserId:       fm.DirectPbwGet[string](req.UserId),
 		PartnerId:    fm.DirectPbwGet[string](req.PartnerId),
@@ -53,13 +50,11 @@ func (stu *stuClog) httpCallV1(req *sclog.RequestHttpCallV1, _ map[string]string
 		ErrorMessage: fm.DirectPbwGet[string](req.ErrMessage),
 		StackTrace:   fm.DirectPbwGet[string](req.StackTrace),
 		Duration:     duration,
+		StartedAt:    fm.GetDefault(startedAt, timenow),
+		FinishedAt:   fm.GetDefault(finishedAt, timenow),
 	}
 
-	e.StartedAt = fm.GetDefault(startedAt, e.StartedAt)
-	e.FinishedAt = fm.GetDefault(finishedAt, e.FinishedAt)
-
-	err = repo.HttpCallV1.Insert(trimHttpCallV1(e))
+	err := repo.HttpCallV1.Insert(trimHttpCallV1(e))
 	saveError(err, e)
-
 	return send(err)
 }
