@@ -8,8 +8,8 @@
 
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v3.21.12
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: res/proto/clog.proto
 
 package sclog
@@ -23,8 +23,8 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
 	CLogService_NoteV1_FullMethodName         = "/sclog.CLogService/NoteV1"
@@ -33,6 +33,7 @@ const (
 	CLogService_ServicePieceV1_FullMethodName = "/sclog.CLogService/ServicePieceV1"
 	CLogService_ServiceV1_FullMethodName      = "/sclog.CLogService/ServiceV1"
 	CLogService_GrpcV1_FullMethodName         = "/sclog.CLogService/GrpcV1"
+	CLogService_DistLockV1_FullMethodName     = "/sclog.CLogService/DistLockV1"
 )
 
 // CLogServiceClient is the client API for CLogService service.
@@ -45,6 +46,7 @@ type CLogServiceClient interface {
 	ServicePieceV1(ctx context.Context, in *RequestServicePieceV1, opts ...grpc.CallOption) (*Response, error)
 	ServiceV1(ctx context.Context, in *RequestServiceV1, opts ...grpc.CallOption) (*Response, error)
 	GrpcV1(ctx context.Context, in *RequestGrpcV1, opts ...grpc.CallOption) (*Response, error)
+	DistLockV1(ctx context.Context, in *RequestDistLockV1, opts ...grpc.CallOption) (*Response, error)
 }
 
 type cLogServiceClient struct {
@@ -56,8 +58,9 @@ func NewCLogServiceClient(cc grpc.ClientConnInterface) CLogServiceClient {
 }
 
 func (c *cLogServiceClient) NoteV1(ctx context.Context, in *RequestNoteV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Response)
-	err := c.cc.Invoke(ctx, CLogService_NoteV1_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CLogService_NoteV1_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -65,8 +68,9 @@ func (c *cLogServiceClient) NoteV1(ctx context.Context, in *RequestNoteV1, opts 
 }
 
 func (c *cLogServiceClient) DbqV1(ctx context.Context, in *RequestDbqV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Response)
-	err := c.cc.Invoke(ctx, CLogService_DbqV1_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CLogService_DbqV1_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -74,8 +78,9 @@ func (c *cLogServiceClient) DbqV1(ctx context.Context, in *RequestDbqV1, opts ..
 }
 
 func (c *cLogServiceClient) HttpCallV1(ctx context.Context, in *RequestHttpCallV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Response)
-	err := c.cc.Invoke(ctx, CLogService_HttpCallV1_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CLogService_HttpCallV1_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -83,8 +88,9 @@ func (c *cLogServiceClient) HttpCallV1(ctx context.Context, in *RequestHttpCallV
 }
 
 func (c *cLogServiceClient) ServicePieceV1(ctx context.Context, in *RequestServicePieceV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Response)
-	err := c.cc.Invoke(ctx, CLogService_ServicePieceV1_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CLogService_ServicePieceV1_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -92,8 +98,9 @@ func (c *cLogServiceClient) ServicePieceV1(ctx context.Context, in *RequestServi
 }
 
 func (c *cLogServiceClient) ServiceV1(ctx context.Context, in *RequestServiceV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Response)
-	err := c.cc.Invoke(ctx, CLogService_ServiceV1_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CLogService_ServiceV1_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -101,8 +108,19 @@ func (c *cLogServiceClient) ServiceV1(ctx context.Context, in *RequestServiceV1,
 }
 
 func (c *cLogServiceClient) GrpcV1(ctx context.Context, in *RequestGrpcV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(Response)
-	err := c.cc.Invoke(ctx, CLogService_GrpcV1_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, CLogService_GrpcV1_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *cLogServiceClient) DistLockV1(ctx context.Context, in *RequestDistLockV1, opts ...grpc.CallOption) (*Response, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(Response)
+	err := c.cc.Invoke(ctx, CLogService_DistLockV1_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -111,7 +129,7 @@ func (c *cLogServiceClient) GrpcV1(ctx context.Context, in *RequestGrpcV1, opts 
 
 // CLogServiceServer is the server API for CLogService service.
 // All implementations must embed UnimplementedCLogServiceServer
-// for forward compatibility
+// for forward compatibility.
 type CLogServiceServer interface {
 	NoteV1(context.Context, *RequestNoteV1) (*Response, error)
 	DbqV1(context.Context, *RequestDbqV1) (*Response, error)
@@ -119,12 +137,16 @@ type CLogServiceServer interface {
 	ServicePieceV1(context.Context, *RequestServicePieceV1) (*Response, error)
 	ServiceV1(context.Context, *RequestServiceV1) (*Response, error)
 	GrpcV1(context.Context, *RequestGrpcV1) (*Response, error)
+	DistLockV1(context.Context, *RequestDistLockV1) (*Response, error)
 	mustEmbedUnimplementedCLogServiceServer()
 }
 
-// UnimplementedCLogServiceServer must be embedded to have forward compatible implementations.
-type UnimplementedCLogServiceServer struct {
-}
+// UnimplementedCLogServiceServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedCLogServiceServer struct{}
 
 func (UnimplementedCLogServiceServer) NoteV1(context.Context, *RequestNoteV1) (*Response, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method NoteV1 not implemented")
@@ -144,7 +166,11 @@ func (UnimplementedCLogServiceServer) ServiceV1(context.Context, *RequestService
 func (UnimplementedCLogServiceServer) GrpcV1(context.Context, *RequestGrpcV1) (*Response, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GrpcV1 not implemented")
 }
+func (UnimplementedCLogServiceServer) DistLockV1(context.Context, *RequestDistLockV1) (*Response, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method DistLockV1 not implemented")
+}
 func (UnimplementedCLogServiceServer) mustEmbedUnimplementedCLogServiceServer() {}
+func (UnimplementedCLogServiceServer) testEmbeddedByValue()                     {}
 
 // UnsafeCLogServiceServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to CLogServiceServer will
@@ -154,6 +180,13 @@ type UnsafeCLogServiceServer interface {
 }
 
 func RegisterCLogServiceServer(s grpc.ServiceRegistrar, srv CLogServiceServer) {
+	// If the following call pancis, it indicates UnimplementedCLogServiceServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&CLogService_ServiceDesc, srv)
 }
 
@@ -265,6 +298,24 @@ func _CLogService_GrpcV1_Handler(srv interface{}, ctx context.Context, dec func(
 	return interceptor(ctx, in, info, handler)
 }
 
+func _CLogService_DistLockV1_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(RequestDistLockV1)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(CLogServiceServer).DistLockV1(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: CLogService_DistLockV1_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(CLogServiceServer).DistLockV1(ctx, req.(*RequestDistLockV1))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 // CLogService_ServiceDesc is the grpc.ServiceDesc for CLogService service.
 // It's only intended for direct use with grpc.RegisterService,
 // and not to be introspected or modified (even as a copy)
@@ -295,6 +346,10 @@ var CLogService_ServiceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GrpcV1",
 			Handler:    _CLogService_GrpcV1_Handler,
+		},
+		{
+			MethodName: "DistLockV1",
+			Handler:    _CLogService_DistLockV1_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},
